Fix -Wsign-compare and -Wmissing-designated-field-initializers for -Wextra
Fix all -Wsign-compare and -Wmissing-designated-field-initializers warnings
in tests so the build compiles cleanly with -Wextra -Werror under Clang 20.

Changes:
- Loop variable type mismatches: change int/uint32_t loop vars to match the
  bound type (size_t for .size()/.x/.y, uint32_t or int for named vars)
- Struct designated initializers: add = std::nullopt / = {} / = 0 defaults
  to optional/vector/numeric fields so partial initializer lists compile

+++ tests/tt_metal/tt_metal/context/test_metal_context_api.cpp
@@ -192,7 +192,7 @@ TEST_F(MetalContextTest, MaxContexts) {
     auto mock_cluster_desc_path = experimental::get_mock_cluster_desc_name(tt::ARCH::WORMHOLE_B0, 1).value();
     std::vector<std::unique_ptr<MetalEnv>> envs;
     envs.reserve(MAX_CONTEXT_COUNT);
-    for (int i = 0; i < MAX_CONTEXT_COUNT - 1; ++i) {
+    for (size_t i = 0; i < MAX_CONTEXT_COUNT - 1; ++i) {
         envs.push_back(std::make_unique<MetalEnv>(MetalEnvDescriptor(mock_cluster_desc_path)));
         context_ids.push_back(MetalContext::create_instance(*envs.back()));
     }
